@@ -1,6 +1,6 @@
 {
   "name": "starterkit-es6",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -11,20 +11,23 @@
   "author": "toabr",
   "license": "ISC",
   "devDependencies": {
+    "@babel/core": "^7.3.3",
+    "autoprefixer": "^9.4.8",
     "babel": "^6.23.0",
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.3",
-    "babel-preset-es2015": "^6.24.1",
-    "css-loader": "^0.28.10",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "i": "^0.3.6",
-    "less": "^2.3.1",
-    "less-loader": "^4.0.6",
-    "rimraf": "^2.6.2",
-    "style-loader": "^0.20.2",
-    "webpack": "^3.1.0",
-    "webpack-dev-server": "^2.11.1",
-    "html-webpack-plugin": "^3.0.4"
+    "babel-core": "^6.26.3",
+    "babel-loader": "^8.0.5",
+    "babel-preset-env": "^1.7.0",
+    "css-loader": "^2.1.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "html-webpack-plugin": "^3.2.0",
+    "node-sass": "^4.11.0",
+    "postcss-loader": "^3.0.0",
+    "rimraf": "^2.6.3",
+    "sass-loader": "^7.1.0",
+    "style-loader": "^0.23.1",
+    "webpack": "^4.29.5",
+    "webpack-cli": "^3.2.3",
+    "webpack-dev-server": "^3.2.0"
   },
   "dependencies": {}
 }
